+++ gjuton-tests/src/test/java/io/github/gjuton/internal/generator/GeneratorContextTest.java
@@ -445,7 +445,8 @@ class GeneratorContextTest {
     }
 
     private static GeneratorContext contextWithOverride(String path, Supplier<Object> override) {
-        var config = new GeneratorConfig(false, false, 2, 4, Map.of(path, override), Map.of(), Map.of(), ValueConstraints.forExhaustive());
+        var config = new GeneratorConfig(false, false, 2, 4, Map.of(path, override), Map.of(), Map.of(),
+                ValueConstraints.forExhaustive(), ValueConstraints.none());
         return new GeneratorContext(new SchemaDocument(new NullSchema(), Map.of()), new Random(1), config);
     }
 }